Normalise steps

+++ srcs/python/kungfu/ops.py
@@ -210,7 +210,7 @@ def start_gpu_group(*args, **kwargs):
     return _op_lib.start_gpu_group(*args, **kwargs)
 
 def _to_gs(epoch, batch_size, num_train):
-    return (epoch * num_train / (batch_size * _get_num_peers()))
+    return int(epoch * num_train / (batch_size * _get_num_peers()))
 
 def to_steps(epochs, batch_size, num_train):
     return [_to_gs(epoch, batch_size, num_train) for epoch in epochs]
